all done just want to clean the code

+++ app.py
@@ -2,7 +2,6 @@ from flask import Flask, render_template, request, redirect, url_for, flash
 from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin, login_user, LoginManager, login_required, logout_user
 from werkzeug.security import check_password_hash
-import json
 import os
 
 app = Flask(__name__)
@@ -35,23 +34,25 @@ class metadata(db.Model, UserMixin):
     photo_path = db.Column(db.String(500))
 
 
-with open('photo_data.json','r') as f:
-    photo_data = json.load(f)
 
 index = [1, 2, 3, 4, 5, 6, 4, 7, 4, 7, 8, 7, 9,2,10,11,12,10,13]
 
 def getting_category():
     get_category = metadata.query.all()
-    unique_category = set([get_category[i].category for i in range(len(get_category))])
+    unique_category = list(set([get_category[i].category for i in range(len(get_category))]))
+    try: 
+        unique_category.remove('')
+    except:
+        pass
     return unique_category
 
 @app.route('/')
 def home():
     metadata_query = metadata.query.all()
-    return render_template('index.html',photo_data=metadata_query,range_len=len(metadata_query),index=index)
+    return render_template('index.html',photo_data=metadata_query,range_len=len(metadata_query),index=index,catergory=getting_category())
 
 @app.route('/dashboard')
-# @login_required
+@login_required
 def dashboard():
     metadata_query = metadata.query.all()
     return render_template('dashboard.html',photo_data=metadata_query,range_len=len(metadata_query),catergory=getting_category(),category_len=len(getting_category()))
@@ -69,7 +70,11 @@ def login():
         if user:
             if check_password_hash(user.password,passwd):
                 login_user(user)
+                flash('Your Are Logged In Successfully')
                 return redirect(url_for('dashboard'))
+            else:
+                flash('Please check your email or Check password is correct or not')
+                return redirect(url_for('login'))
         else:
             flash('Please check your email or Check password is correct or not')
             return redirect(url_for('login'))
@@ -78,13 +83,15 @@ def login():
 @login_required
 def logout():
     logout_user()
+    flash('You Are Logged Out Successfully')
     return redirect('/')
 
 
 @app.route('/upload', methods=['GET','POST'])
+@login_required
 def upload():
     if request.method == 'GET':
-        return render_template('upload.html',catergory=getting_category() )
+        return render_template('upload.html',catergory=getting_category())
     else:
         title = request.form['title']
         sub_title = request.form['sub_title']
@@ -108,24 +115,21 @@ def upload():
 @app.route('/update',methods=['POST'])
 def update():
     if request.method == 'POST':
-
-
         title = request.form['title']
         sub_title = request.form['sub_title']
         category = request.form['category']
-        id = request.form['ikivalue']       
+        id = request.form['ikivalue']
         photo_data = metadata.query.filter_by(id=id)
-        try:
-            path = request.files['file']
+        path = request.files['file']
+        if path.filename != '':
             path1 = os.path.join(app.config['UPLOAD_FOLDER'],path.filename)
             path.save(path1)
             os.remove(f'static/{photo_data[0].photo_path}')
-        except FileNotFoundError:
-            pass
 
-        metadata.query.filter_by(id=id).update(dict(title=title if title !='' else photo_data[0].title,sub_title=sub_title if sub_title !='' else photo_data[0].sub_title,category=category if category !='' else photo_data[0].category,photo_path=path1.replace('static/','') if path !='' else photo_data[0].photo_path))
+        metadata.query.filter_by(id=id).update(dict(title=title if title !='' else photo_data[0].title,sub_title=sub_title if sub_title !='' else photo_data[0].sub_title,category=category if category !='' else photo_data[0].category,
+        photo_path=path1.replace('static/','') if path.filename != '' else photo_data[0].photo_path))
         db.session.commit()
-
+        flash('Detail Updated')
         return redirect(url_for('dashboard'))
         
 @app.route('/delete',methods=['POST'])
@@ -135,11 +139,23 @@ def delete():
         delete_data = metadata.query.filter_by(id=id).first()
         try:
             os.remove(f'static/{delete_data.photo_path}')
-        except FileNotFoundError:
+        except :
             pass
         db.session.delete(delete_data)
         db.session.commit()
+        flash('The Selected Photo Has Been Deleted!')
         return redirect(url_for('dashboard'))
+
+
+@app.route('/delete-category',methods=['POST'])
+def delete_category():
+    if request.method == 'POST':
+        delete_category = request.form['category']
+        metadata.query.filter_by(category=delete_category).update(dict(category=''))
+        db.session.commit()
+        flash(f'{delete_category} delete successfully')
+        return redirect('dashboard')
+
 
 if __name__ == '__main__':
     app.run(debug=True,host='192.168.0.107')
